requirements.txt
attrs==23.1.0
bracex==2.3.post1
cached-property==1.5.2
Cerberus==1.3.4
certifi==2023.5.7
charset-normalizer==3.1.0
coverage==7.2.7
ecs-logging==1.1.0
elastic-enterprise-search==8.4.0
elastic-transport==8.4.0
elasticsearch==8.7.0
flake8==4.0.1
idna==3.4
iniconfig==2.0.0
mccabe==0.6.1
packaging==23.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.5.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.7.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-custom-exit-code==0.3.0
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
tika==1.24
toml==0.10.2
tomli==2.0.1
urllib3==1.26.16
wcmatch==8.3
pymysql==1.1.0
ldap3==2.9.1
leadtools==22.0.0.8
google-api-python-client==2.103.0
facebook-sdk==3.1.0